README.md
Hi! My name is Darsh Patel, and I am currently a student at University of Maryland Baltimore County, pursuing a Bachelor's degree in Computer Science, and I am in my second year at the school. 
In this Github, you will find a link to some of the most challenging assignments I have worked on throughout my time in University, as I emerge myself into the world of coding, with an aim to eventually find myself doing something in the field of Cybsecurity.
My phone number is: 443-840-0170
My personal email is: darshp717@gmail.com